ENGR00289053-1 mx6sabresd: Fix wrong colors in LVDS display
Currently HDMI splash screen is selected by default on mx6sabresd boards.

As LVDS is also always enabled, this causes incorrect colors to be displayed
in the LVDS panel due to the different pixel format in HDMI and LVDS.

Fix this by selecting the LVDS panel as the default splash output and also by
ensuring that LVDS is turned off when HDMI output is active.

Signed-off-by: Fabio Estevam <fabio.estevam@freescale.com>
Signed-off-by: Liu Ying <Ying.Liu@freescale.com>