ENGR00288995 MX6 SabreSD: Correct IPU DI pix fmt for LDB
The DC mapping for LVDS666 is different from that for
RGB666. Currently, we set IPU DI pixel format to be
LVDS666 and set LDB data width to be 24bit. This happens
to make the display work normally somehow. But, the two
configurations are wrong and don't match with each other.
This patch corrects the IPU DI output pixel format from
LVDS666 to RGB666 and LDB data width from 24bit to 18bit.

Signed-off-by: Liu Ying <Ying.Liu@freescale.com>