ENGR00294905 driver:MAX7310: Add GPIO expander driver for MAX7310
Implement simple functionalities for MAX7310 GPIO input and output.
Because MAX7310 is a off-chip device and need to co-exist with
on-chip GPIO, new APIs are added specifically for expander device.

CONFIG_MAX7310_IOEXP is used to enable the MAX7310 driver. The I2C
related configurations also need to set together.

Signed-off-by: Ye.Li <B37916@freescale.com>