ENGR00300374 video: ipu: wait for DP SF end irq when disabling sync BG flow
Instead of waiting for DC triple buffer to be cleared, this patch changes
to wait for a relevant DP sync flow end irq when disabling sync BG flows.
In this way, we align the implement to the FSL internal IPUv3 driver.
After applying this patch, the uboot hang up issue at the arch_preboot_os
stage on the MX6DL platforms is not observed any more.

Signed-off-by: Liu Ying <Ying.Liu@freescale.com>