ENGR00301683 FEC: Modify enet driver TX polling bit
Change to check READY bit in BD, not check the TDAR. On iMX6SX, FEC will
clear the TDAR prior than the READY bit of last BD. Since fec driver only
prepare two BD for transmit, this cause the BD send failed at the third
packet.

On iMX6SX, we disabled DCACHE to workaround for this issue. Now change to
enable the DCACHE.

Signed-off-by: Ye.Li <B37916@freescale.com>