ENGR00315894-32 i.mx6:sabreauto/sabresd Set the mmcdev for various boot devices
For the SPI/WEIM NOR, it can't store all the images into it due to it's
capacity, we need one default mmc device to load the left image or rootfs.
The end user need change the default setting according to their needs.
For NAND/SATA boot, the storage is big enough to hold all the stuff.
For SD/MMC boot, mmcdev is dynamiclly created due to the boot SD/MMC slot.

Signed-off-by: Jason Liu <r64343@freescale.com>
Signed-off-by: Ye.Li <B37916@freescale.com>