ENGR00315894-65 iMX6SX:SABRESD Add RDC settings to BSP
According to the SRS, in the M4 CAN demo, the GPIO group1 will be
shared between A9 and M4. At A9 side, the pins 0, 1, 2, 3 are used.
M4 also uses one pin in its application.

To synchronize the registers setttings of GPIO1, must enable RDC
and RDC semaphore on the GPIO1.

Signed-off-by: Ye.Li <B37916@freescale.com>