ENGR00329930-2 MMC: Fix debounce problem in the tuning algorithm
The boundary value for the tuning is not stable. It is not always can
tune successfully. Current algorithm picks up the boundary value as min
vertex, when starting finding the max vertex, the bounday value may
fail and set the max vertex same to min. So a incorrect tuning value is
got, and causes the SD card failed.

Change the tuning algorithm to select a max window that every value in
the window can tune successfully. So the debounce problem is fixed.

Signed-off-by: Ye.Li <B37916@freescale.com>
(cherry picked from commit b36fd693586dc70191b1d684f04410bd61325477)