MLK-9706 imx: mx6sx19x19arm2: Fix ENET card MAX7322 reset issue
The MAX7322 will fail to work on 19x19 arm2 revB board. This failure
is caused by the MAX7322 reset pin is not released when calling the
setup_fec function.

The MAX7322 reset pin is same as PHY reset pin. This patch fixes the issue
by moving the PHY reset from setup_iomux_fec1 to setup_fec.

Signed-off-by: Ye.Li <B37916@freescale.com>