MLK-10360 imx: iomux-v3: Fix the driver issue for MUX_CTL with offset 0
The IOMUX-v3 driver checks the MUX_CTL register offset. If the
offset is zero, it will skip the MUX_CTL register setting.

This behavior is correct for previous platforms like i.MX6,
but on i.MX7, the IOMUXC_LPSR_SW_MUX_CTL_PAD_GPIO1_IO00 is at offset 0.
Thus, it causes the MUX setting to this pin always not working.

This fix removes this condition checking since it is unnecessary.

Signed-off-by: Ye.Li <B37916@freescale.com>