MLK-10733 imx: imx6qpsabreauto: Update IPU QoS settings
Update IPU QoS settings from 0x007f007f to 0x77177717
according to the SoC team's recommendation.  This change
should be able to balance AXI ID0/2/3 priority and set
AXI ID1 priority relatively lower, which matches the way
we use AXI ID0/1/2/3 for IDMAC23(0), regular IDMACs,
IDMAC27 and IDMAC28 respectively in kernel.  The specific
priority values for each AXI ID are supposed to be picked
for the sake of an overall good system performance.

Signed-off-by: Liu Ying <Ying.Liu@freescale.com>