MLK-11795-02 imx: upate the pmic standby voltage for imx6qp
According to the latest datasheet(Rev. B, 07/2015), the VDD_SOC_IN
standby voltage should be 1.05V and on i.MX6QP, we can use the PMIC
'APS' mode in standby. we add a 25mV margin to cover the IR drop and
board tolerance, so the standby voltage of VDD_SOC_IN should be
setting to 1.075V.

Signed-off-by: Bai Ping <b51503@freescale.com>