MLK-11823 USB:gadget Fix USB port interface issue in ci_udc driver
The ci_udc driver tries to use the ULPI interface for the USB OTG controller,
but this type is not supported by all i.MX6 and i.MX7 platforms. When setting to
ULPI, other platforms except the 6UL refuse the settings and keep default value.
But on 6UL, the PTW bit of PORTSC1 register which is documented as RO can change.
This cause the interface setting problem with USB PHY.

Fix the issue by removing the ULPI setting for i.MX6 and i.MX7. All will use default
UTMI setting.

Signed-off-by: Ye.Li <B37916@freescale.com>