s32v234: qspi: add commands to manipulate hyperflash
These are simple custom commands to allow access to the hyperflash,
but in the future we should use the standard u-boot commands provided
via CONFIG_CMD_FLASH interface. For now our simple commands are enough
for basic functionality.

The proper way to protect the flash from accidental writes and erases
would be through the HW mechanisms of the flash, but we still have
to add some basic functionality, so HW protect will be inplemented later.

For now we simply add a software flag to prevent accidental erases and
writes and we make sure that when the protection is on, we clean up the
LUTs which we use for the erase and write commands. This is not ideal,
but should work for now.

Commands:

    protect on - enables SW protection of flash

    protect off - allows erase and write operations on flash

This means we have a namespace collision on the flash related commands
(e.g. erase, protect), so we make sure we override the default flash
commands with our own versions and we #undef CONFIG_CMD_FLASH to aviod
collisions.

Signed-off-by: Eddy Petrișor <eddy.petrisor@nxp.com>
Signed-off-by: Larisa Grigore <Larisa.Grigore@nxp.com>