s32v234: DDR handshake at functional reset
In order for the DDR to be stabilized after a functional reset,
DDR handshake feature was implemented and enabled (any reset
functional event will not be propagated to the device until DDR
is moved to a safe state).

After a functional reset, before configuring the DDR, it is checked
if the handshake was done and if the DDR is in a safe state.

The handshake shouldn't be done at the first boot.
Otherwise u-boot can't be downloaded from a debugger which triggers an
external reset because the handshake can't be completed.

Signed-off-by: Ana Nedelcu <ana.nedelcu@nxp.com>
Signed-off-by: Cosmin-Stefan Stoica <cosmin.stoica@nxp.com>
Signed-off-by: Larisa Grigore <Larisa.Grigore@nxp.com>