MLK-12575-2 imx: Modify environment offset to 6 sectors on EIMNOR
On mx6qsabreauto and mx6sx ARM2 boards, the EIMNOR sector size is 128KB.
And its u-boot environment offset is 4 sectors (512KB). But u-boot size has
exceeds it, so change to 6 sectors offset (768KB).
To align the environment configurations for all i.MX, also change the configuration
for mx6ul and mx7d, which has EIMNOR with 256KB sectors.

Signed-off-by: Ye Li <ye.li@nxp.com>