MLK-12723 imx: Change the env offset on NAND to 60M
Current environment offset on NAND is 37MB, this will cause a alignment
issue when erasing if nand erase block is 2MB. The saveenv is failed.

=> saveenv
Saving Environment to NAND...
Erasing NAND...
Attempt to erase non block-aligned data

Since the max erase block we supported is 4MB, adjust the env offset to 60MB,
where is the last 4MB in 64MB reserved area for boot.

Signed-off-by: Ye Li <ye.li@nxp.com>