s32v234: clk: Ungate the clock only for the IPS
The Periperal configuration was moved from RUN_PC0 to RUN_PC1,
because when RUN_PC0 is enabled for periperals, all
peripherals have clock by default. In order to avoid this situation,
RUN_PC0 is kept disabled and RUN_PC1 is enabled. The Periperals which
require clock are set to work under RUN_PC1.

Signed-off-by: Stoica Cosmin-Stefan <cosmin.stoica@nxp.com>
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>