MXSCM-213: remove dmfc from boot arguments for mx6scm boards
Intially this parameter was added to fix a video stuttering but
with L4.1 the video issue is not present so we can safely get rid
of this parameter.

When using both ldb interfaces in separate mode and passing the
dmfc argument as boot parameter to the kernel, a distortion on
both displays is observed when rendering to the secondary display.

Signed-off-by: Juan Gutierrez <juan.gutierrez@nxp.com>