MLK-13450-3 mx7ulp: Add iomux driver to support IOMUXC0 and IOMUXC1
Add a new driver under ULP directory to support its IOMUXC controllers.
The ULP has two IOMUXC, the IOMUXC0 is used for M4 domain, while IOMUXC1 is
for A7. We set IOMUXC1 as the default IOMUX in this driver. Any pins in
IOMUXC0 needs to configure with IOMUX_CONFIG_MPORTS in its mux_mode field.

Signed-off-by: Ye Li <ye.li@nxp.com>
Signed-off-by: Shenwei Wang <shenwei.wang@nxp.com>