MLK-13450-12 gpio: Add Rapid GPIO2P driver for i.MX7ULP
Add the imx_rgpio2p driver for Rapid GPIO2P controllers on i.MX7ULP. Have added
all ports on RGPIO2P_0 and RGPIO2P_1.

The configurations CONFIG_IMX_RGPIO2P and CONFIG_DM_GPIO must be set to y to
enable the drivers.

To use the GPIO function, the IBE and OBE needs to set in IOMUXC. We did
not set the bits in driver, but leave them to IOMUXC settings of the GPIO pins.
User should use IMX_GPIO_NR to generate the GPIO number for gpio APIs access.

Signed-off-by: Ye Li <ye.li@nxp.com>