MXSCM-222 usb: fix swbst_mode to properly power usb
The definition of the SWBST_MODE_AUTO at the pfuze100_pmic.h file
changed between uboot versions. On the previous version the shift
to the proper bit field was part of the macro. In the uboot v2016
this macro does not include the shift and needs to be performed
explicitly to properly modify the SWBST_MODE bit field.

Signed-off-by: Juan Gutierrez <juan.gutierrez@nxp.com>