MLK-14328 imx6slevkandroid: Remove config.h from mxc_key_defs.h header file
The mx6slevk android build is broken after adding CONFIG_EFI_PARTITION.
The CONFIG_LIB_UUID is not set which should be set by "config_fallbacks.h" when
having CONFIG_EFI_PARTITION defined.

The root cause is in "mx6slevkandroid.h", we include the "mxc_key_defs.h" which
includes the "config.h". Since the "config.h" includes the "config_fallbacks.h",
so the "config_fallbacks.h" is included twice in building. The first time is
before the CONFIG_EFI_PARTITION definition, the second is after it.

Because "config_fallbacks.h" uses __CONFIG_FALLBACKS_H to avoid being included multiple
times. So the second is not effective and cause CONFIG_LIB_UUID is finally not set.

This patch removes the "config.h" from "mxc_key_defs.h", no need to include it.

Signed-off-by: Ye Li <ye.li@nxp.com>