MLK-14484-1 mtd/spi: Add MT35XU512ABA NOR flash support
Add MT35XU512ABA parameters to NOR flash parameters array. Since the
manufactory ID is changed to 0x2C, add it for micron and using it for
relevant settings.

The MT35XU512ABA only supports 1 bit mode and 8 bits. It can't support
dual and quad. Because the 8 bits is not support by u-boot framework and
driver. We only use 1 bit mode for this flash.

Signed-off-by: Ye Li <ye.li@nxp.com>