MLK-18732 imx8: Add fuse driver for read/write OTP memory
Implement a fuse driver to access OTP memory. When CONFIG_SMC_FUSE is set,
the driver uses SIP call to ATF to read/write OTP. When CONFIG_SMC_FUSE is
not set, the driver wraps the SCFW API and only provide interfaces
to read the OTP.

Since there is no concept of fuse bank on i.MX8. Need set "bank" parameter to 0
when using the fuse command.

Signed-off-by: Ye Li <ye.li@nxp.com>
Signed-off-by: Teo Hall <teo.hall@nxp.com>