MA-10341 Rename physical partition boot0 from bootloader to bootloader0.
On the imx7d, the bootloader currently in use resides on physical partition
boot0, and is currently referenced and hard coded as "bootloader".  But
this conflicts with the new bootloader_a and bootloader_b in the gpt, so
that "fastboot flash bootloader ..." flashes to bootloader_a or b. With this
fix boot0 can instead by explicitly referenced as
"fastboot flash bootloader0 ...".

"bootloader0" is currently the only bootloader on the system, but
once a/b bootloading is implemented it is going to be a more SPL-like
that loads either bootloader_a or bootloader_b.

Test: manually verified flashing bootloader0, bootloader, and bootloader_a
      and bootloader_b.

Change-Id: I03a85289f80cdf5f1b80dc07a962a143c6a351e1