[iot] fastboot:Add 'get_staged' command support
When host end execute "fastboot get_staged [ <outfile> ]"
The staged data from the last command will be output to host end.
If the last command did not result in staged data or the target data
larger than end point buffer size, this command will be fail.

This patch also alloc new buffer that is larger than normal end point's
as the this command need to transfer data by one packet. So that
only the file that less than 128KB will be transfered.

Change-Id: I1dbefa130df9ca348a30d7bc52cb856c052776bf
Signed-off-by: Haoran Wang <elven.wang@nxp.com>